requirements.txt
Flask==1.1.1
h5py==2.10.0
imutils==0.5.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
numba==0.47.0
numpy==1.17.5
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
pandas==1.0.1
Pillow==6.2.2
scikit-learn==0.22.1
scipy==1.4.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tqdm==4.42.1
Werkzeug==0.16.0
